refactor(internal): extract errors → internal/errs and log → internal/slogx (compat-preserving)
Phase 1 of the v2 restructure plan. Pure internal refactor — public
API is byte-for-byte unchanged (verified by the
TestErrAliasChain_v1Compat test in errors_test.go and the existing
integration suite running green against the new layout).

Why now: Phase 2 (move convert_*.go to convert/ subpackage) and
Phase 3 (move publish-pipeline files to publish/ subpackage) both
need to call newGISError and the default *slog.Logger. If those
symbols stay in the root gismanager package, the new subpackages
have to import the root package — which then can't import them
back without creating a circular import. Pulling errors and the
logger to internal/ subpackages breaks that future cycle before
it forms.

What moved:

- errors.go content → internal/errs/errs.go. Function rename
  newGISError → NewGISError (mandatory: cross-package access
  needs an exported name). Sentinels and the *GISError type
  keep their original names.
- log.go GetLogger() → internal/slogx/Default(). Renamed in the
  internal package because Default is the more idiomatic name
  inside a slogx package.

What stayed (v1.x compat layer):

- errors.go still declares `type GISError = errs.GISError` and
  `var ErrConfigInvalid = errs.ErrConfigInvalid` (etc.) so
  external callers continue to use gismanager.GISError /
  gismanager.ErrConvertFailed unchanged. Type aliases preserve
  full identity — errors.As / errors.Is / pointer equality work
  transparently across the boundary.
- errors.go also keeps a package-private `newGISError` wrapper
  delegating to errs.NewGISError. Every existing caller in
  manager.go / layer.go / walk.go / utils.go / convert_*.go is
  byte-identical — they keep calling newGISError(...) without
  knowing about the move. Phase 2+ moves those callers to a
  subpackage where they call errs.NewGISError directly; Phase 4
  drops the wrapper entirely.
- log.go keeps GetLogger() as a thin wrapper around
  slogx.Default() for v1.x compat.

Tests:

- internal/errs/errs_test.go gets the comprehensive *GISError
  mechanic tests (TestGISError_IsAndUnwrap / TestGISError_As /
  TestGISError_NestedChain / TestGISError_JoinedAggregation /
  TestGISError_ErrorFormatsByCombination), now using the exported
  NewGISError name. Provides direct package coverage.
- internal/slogx/default_test.go gets a small two-test suite
  (Default returns non-nil; Default returns distinct instances
  per call so callers can WithAttrs without cross-talk).
- errors_test.go (root) shrinks to a single
  TestErrAliasChain_v1Compat test that confirms the alias
  resolves to the same instances as internal/errs — guards
  against a future refactor accidentally breaking the alias
  boundary for external callers.
- log_test.go (root) keeps the existing TestGetLogger.

Full unit suite + integration suite (GeoServer 2.28.0, ~15s
publish-pipeline runtime) green inside the dev container. No
external API changes; existing publish_integration_test.go +
convert_*_integration_test.go cases unaffected.
