+++ log.go
@@ -2,17 +2,20 @@ package gismanager
 
 import (
 	"log/slog"
-	"os"
+
+	"github.com/hishamkaram/gismanager/internal/slogx"
 )
 
 // GetLogger returns the project's default structured logger: a stdlib
 // *slog.Logger writing text-formatted records to stderr.
 //
-// Callers building their own pipeline should construct an *slog.Logger
-// directly (any handler — JSON, lumberjack-rotated, otel, etc.) and pass
-// it on the [ManagerConfig].logger field via the constructor (PR 4 still
-// loads it from FromConfig + GetLogger; functional-options constructor
-// lands in a follow-up).
+// As of the v2 restructure groundwork (Phase 1), the implementation
+// lives in [internal/slogx.Default]; this is a thin wrapper for
+// v1.x compatibility. v2 callers should construct their own
+// *slog.Logger directly (any handler — JSON, lumberjack-rotated,
+// OpenTelemetry-bridged) and pass it on via the available
+// functional options. The internal package is not part of the
+// v2 public API.
 func GetLogger() *slog.Logger {
-	return slog.New(slog.NewTextHandler(os.Stderr, nil))
+	return slogx.Default()
 }
